+++ model_navigator/model_analyzer/analyzer.py
@@ -145,7 +145,7 @@ class AnalysisConfigGenerator(BaseConfigGenerator):
         for model_name in model_names:
             LOGGER.info(f"\t- {model_name}")
 
-        # https://github.com/triton-inference-server/model_analyzer/blob/r22.01/docs/config.md
+        # https://github.com/triton-inference-server/model_analyzer/blob/r22.02/docs/config.md
         config = {
             "analysis_models": model_names,
             "checkpoint_directory": self._analyzer_checkpoints_dir_path.as_posix(),