+++ CHANGELOG.md
@@ -16,6 +16,24 @@ limitations under the License.
 
 # Changelog
 
+## 0.2.7
+- Updated NVIDIA containers defaults to 22.02
+- Removed support for Python 3.7
+- Triton Model configuration related:
+  - Support dynamic batching without setting preferred batch size value
+- Profiling related:
+  - Deprecated `--config-search-max-preferred-batch-size` flag as is no longer supported in Triton Model Analyzer
+
+[//]: <> (put here on external component update with short summary what change or link to changelog)
+- Version of external components used during testing:
+    - [Polygraphy](https://github.com/NVIDIA/TensorRT/tree/master/tools/Polygraphy/): 0.35.1
+    - [GraphSurgeon](https://github.com/NVIDIA/TensorRT/tree/master/tools/onnx-graphsurgeon/): 0.3.14
+    - [Triton Model Analyzer 1.8.2](https://github.com/triton-inference-server/model_analyzer)
+    - tf2onnx: [v1.9.3](https://github.com/onnx/tensorflow-onnx/releases/tag/v1.9.3)
+    - Other component versions depend on the used framework and Triton Inference Server containers versions.
+      See its [support matrix](https://docs.nvidia.com/deeplearning/frameworks/support-matrix/index.html)
+      for a detailed summary.
+
 ## 0.2.6
 - Updated NVIDIA containers defaults to 22.01
 - Removed support for Python 3.6 due to EOL